requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Medical Image Processing
monai[all]>=1.3.0
nibabel>=5.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: Visualization
matplotlib>=3.7.0
